refactor: merge integration tests
